\name{BioImageDbs-package}
\alias{BioImageDbs-package}
\alias{BioImageDbs}
\docType{package}
\title{
\packageTitle{BioImageDbs}
}
\description{
\packageDescription{BioImageDbs}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{BioImageDbs}
\packageIndices{BioImageDbs}
}
\author{

\packageAuthor{BioImageDbs}

Maintainer: \packageMaintainer{BioImageDbs}
}

\keyword{ package }
\examples{
ls("package:BioImageDbs") # This is empty package
}
